Fix IEP template routes: use enrollment-aware auth for teacher-student access
The IEP template apply/recommend routes and teacher IEP GET routes were
checking only `teacherId` on the User model, which missed students linked
via enrollment codes (e.g., Ms. Cooper's IEP intervention group). Extracted
`getStudentIdsForTeacher` to a shared service and updated all routes to
check both direct assignment and enrollment code membership.
