+++ routes/teacher.js
@@ -13,31 +13,7 @@ const ScreenerSession = require('../models/screenerSession');
 const EnrollmentCode = require('../models/enrollmentCode');
 const Skill = require('../models/skill');
 const { callLLMStream } = require('../utils/openaiClient');
-
-/**
- * Get all student IDs for a teacher, combining:
- *  1. Students with teacherId pointing to this teacher
- *  2. Students enrolled via this teacher's enrollment codes
- * Returns deduplicated ObjectId array.
- */
-async function getStudentIdsForTeacher(teacherId) {
-  // 1. Direct teacherId assignment
-  const directStudents = await User.find(
-    { role: 'student', teacherId },
-    '_id'
-  ).lean();
-  const idSet = new Set(directStudents.map(s => s._id.toString()));
-
-  // 2. Students enrolled via this teacher's enrollment codes
-  const codes = await EnrollmentCode.find({ teacherId }, 'enrolledStudents').lean();
-  for (const code of codes) {
-    for (const e of (code.enrolledStudents || [])) {
-      idSet.add(e.studentId.toString());
-    }
-  }
-
-  return [...idSet];
-}
+const { getStudentIdsForTeacher } = require('../services/userService');
 
 // Fetches students assigned to the logged-in teacher (via teacherId OR enrollment codes)
 router.get('/students', isTeacher, async (req, res) => {
@@ -70,10 +46,15 @@ router.get('/students/:studentId/iep', isTeacher, async (req, res) => {
     const { studentId } = req.params;
     const teacherId = req.user._id;
 
+    // Check both direct assignment AND enrollment code assignment
+    const authorizedStudentIds = await getStudentIdsForTeacher(teacherId);
+    if (!authorizedStudentIds.includes(studentId)) {
+      return res.status(404).json({ message: 'Student not found or not assigned to this teacher.' });
+    }
+
     const student = await User.findOne({
       _id: studentId,
-      role: 'student',
-      teacherId: teacherId
+      role: 'student'
     }, 'firstName lastName username iepPlan').lean();
 
     if (!student) {
@@ -123,10 +104,15 @@ router.get('/students/:studentId/iep/goal-history', isTeacher, async (req, res) 
     const { studentId } = req.params;
     const teacherId = req.user._id;
 
+    // Check both direct assignment AND enrollment code assignment
+    const authorizedStudentIds = await getStudentIdsForTeacher(teacherId);
+    if (!authorizedStudentIds.includes(studentId)) {
+      return res.status(404).json({ message: 'Student not found or not assigned to this teacher.' });
+    }
+
     const student = await User.findOne({
       _id: studentId,
-      role: 'student',
-      teacherId: teacherId
+      role: 'student'
     }, 'firstName lastName iepPlan').lean();
 
     if (!student) {
